カメラ精度とユーザーインタラクションの改善
主要な改善点:
- カメラ位置精度向上: エクストリンシクスデータをfloatからdouble精度にアップグレード
- 黒点フィルタリング追加: カラーデータが無効な箇所のアーティファクトを除去
- マウス軌道カメラの全面的な書き直し: 直感的な操作（左クリック=回転、中クリック=パン、右クリック=リセット、スクロール=ズーム）
- configuration.yamlからの深度バイアス補正サポート追加
- デバッグログの改善: 6桁精度表示でトラブルシューティングを向上

技術的変更:
- ExtrinsicsLoader: より高精度なList<double>に変更
- DepthMeshGenerator: RGB閾値>5での黒点フィルタリング追加
- MouseOrbitCamera: スムーズ補間、パン機能、フレームレート非依存制御追加
- BinaryDataParser: 深度バイアス読み込みとデバッグ出力フォーマット強化

これによりカメラアライメント問題が解決され、3D復元のナビゲーション体験が大幅に向上しました。

+++ Assets/Script/datapaser/ExtrinsicsLoader.cs
@@ -44,8 +44,8 @@ public class ExtrinsicsLoader
 
 
         // ✅ Xのみ反転。Zは反転しない
-        position = new Vector3(t[0], t[1], t[2]);
-        rotation = new Quaternion(q[0], q[1], q[2], q[3]);
+        position = new Vector3((float)t[0], (float)t[1], (float)t[2]);
+        rotation = new Quaternion((float)q[0], (float)q[1], (float)q[2], (float)q[3]);
 
         return true;
     }
@@ -62,8 +62,8 @@ public class ExtrinsicsLoader
         var q = device.colorCamera_q_depthCamera;
         if (t == null || q == null || t.Count != 3 || q.Count != 4) return false;
 
-        position = new Vector3(t[0], t[1], t[2]);
-        rotation = new Quaternion(q[0], q[1], q[2], q[3]);
+        position = new Vector3((float)t[0], (float)t[1], (float)t[2]);
+        rotation = new Quaternion((float)q[0], (float)q[1], (float)q[2], (float)q[3]);
 
         return true;
     }
@@ -89,9 +89,9 @@ public class ExtrinsicsDevice
 {
     public string serialNumber { get; set; }
     public float depthScaleFactor { get; set; }
-    public List<float> global_t_colorCamera { get; set; }
-    public List<float> global_q_colorCamera { get; set; }
+    public List<double> global_t_colorCamera { get; set; }
+    public List<double> global_q_colorCamera { get; set; }
 
-    public List<float> colorCamera_t_depthCamera { get; set; }
-    public List<float> colorCamera_q_depthCamera { get; set; } 
+    public List<double> colorCamera_t_depthCamera { get; set; }
+    public List<double> colorCamera_q_depthCamera { get; set; } 
 }
